# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Pazaryeri API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: List listings with filters
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: category
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: integer }
        - name: maxPrice
          in: query
          schema: { type: integer }
        - name: city
          in: query
          schema: { type: string }
        - name: sellerId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Paginated listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"
    post:
      operationId: createListing
      tags: [listings]
      security: [{ bearerAuth: [] }]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /listings/me:
    get:
      operationId: getMyListings
      tags: [listings]
      security: [{ bearerAuth: [] }]
      summary: Get current user's listings
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: User listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /listings/{listingId}:
    get:
      operationId: getListingById
      tags: [listings]
      summary: Get listing detail
      parameters:
        - $ref: "#/components/parameters/listingId"
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateListing
      tags: [listings]
      security: [{ bearerAuth: [] }]
      summary: Update listing
      parameters:
        - $ref: "#/components/parameters/listingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingRequest"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteListing
      tags: [listings]
      security: [{ bearerAuth: [] }]
      summary: Delete listing
      parameters:
        - $ref: "#/components/parameters/listingId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /listings/{listingId}/status:
    patch:
      operationId: updateListingStatus
      tags: [listings]
      security: [{ bearerAuth: [] }]
      summary: Update listing status
      parameters:
        - $ref: "#/components/parameters/listingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      security: [{ bearerAuth: [] }]
      summary: List user favorites
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Favorites list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /search:
    get:
      operationId: searchListings
      tags: [search]
      summary: Search listings
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: category
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: integer }
        - name: maxPrice
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"
  /reports:
    post:
      operationId: createReport
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Report listing or user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportRequest"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /blocks/{userId}:
    post:
      operationId: blockUser
      tags: [users]
      security: [{ bearerAuth: [] }]
      summary: Block a user
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: Blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: unblockUser
      tags: [users]
      security: [{ bearerAuth: [] }]
      summary: Unblock a user
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: Unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      security: [{ bearerAuth: [] }]
      summary: List notifications
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      security: [{ bearerAuth: [] }]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
    listingId:
      name: listingId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
    UserProfile:
      type: object
      required: [id, name, rating, totalSales, isVerified, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        name: { type: string }
        avatar: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        city: { type: string, nullable: true }
        district: { type: string, nullable: true }
        rating: { type: number }
        totalSales: { type: number }
        isVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    ListingSummary:
      type: object
      required: [id, title, price, category, status, views, createdAt, image, seller]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        price: { type: integer }
        category: { type: string }
        status: { type: string, enum: [active, sold, reserved, deleted] }
        city: { type: string, nullable: true }
        district: { type: string, nullable: true }
        location: { type: string, nullable: true }
        views: { type: integer }
        isFavorite: { type: boolean }
        distance: { type: string, nullable: true }
        image: { type: string }
        createdAt: { type: string, format: date-time }
        seller:
          type: object
          required: [id, name, avatar, isVerified]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            avatar: { type: string, nullable: true }
            isVerified: { type: boolean }
    ListingDetail:
      allOf:
        - $ref: "#/components/schemas/ListingSummary"
        - type: object
          required: [description, images, acceptsOffers, sellerId]
          properties:
            description: { type: string }
            images:
              type: array
              items: { type: string }
            acceptsOffers: { type: boolean }
            sellerId: { type: string, format: uuid }
            latitude: { type: number, nullable: true }
            longitude: { type: number, nullable: true }
            seller:
              $ref: "#/components/schemas/UserProfile"
    ListingListResponse:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListingSummary"
        nextCursor: { type: string, nullable: true }
        hasMore: { type: boolean }
    CreateListingRequest:
      type: object
      required: [title, price, category, description, images]
      properties:
        title: { type: string, minLength: 3, maxLength: 200 }
        price: { type: integer, minimum: 0 }
        category: { type: string }
        description: { type: string }
        city: { type: string }
        district: { type: string }
        location: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        acceptsOffers: { type: boolean, default: true }
        images:
          type: array
          items: { type: string }
          minItems: 1
          maxItems: 10
    UpdateListingRequest:
      type: object
      properties:
        title: { type: string }
        price: { type: integer }
        category: { type: string }
        description: { type: string }
        city: { type: string }
        district: { type: string }
        location: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        acceptsOffers: { type: boolean }
        images:
          type: array
          items: { type: string }
    UpdateListingStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [active, sold, reserved, deleted] }
...
    CreateReportRequest:
      type: object
      required: [reason]
      properties:
        listingId: { type: string, format: uuid }
        reportedUserId: { type: string, format: uuid }
        reason: { type: string }
        description: { type: string }
    Notification:
      type: object
      required: [id, type, title, body, isRead, createdAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        title: { type: string }
        body: { type: string }
        data: { type: string, nullable: true }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
    NotificationListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Notification"